requirements.txt
typer==0.7.0
python-dotenv==0.21.0
requests==2.28.1
rich
anyio==3.6.2